+++ dht-root/dht-common/src/main/java/com/retailers/dht/common/vo/GoodsGdcprelVo.java
@@ -7,7 +7,7 @@ import com.retailers.dht.common.entity.GoodsGdcprel;
  */
 public class GoodsGdcprelVo extends GoodsGdcprel {
     private String gsName;
-    private Float gdPrice;
+    private Long gdPrice;
     private Long gdResidueinventory;
     private Long sumcount;
     private String imgurl;
@@ -21,11 +21,11 @@ public class GoodsGdcprelVo extends GoodsGdcprel {
         this.gsName = gsName;
     }
 
-    public Float getGdPrice() {
+    public Long getGdPrice() {
         return gdPrice;
     }
 
-    public void setGdPrice(Float gdPrice) {
+    public void setGdPrice(Long gdPrice) {
         this.gdPrice = gdPrice;
     }
